Data.Primitive.PVar

Description

Documentation

PVar has significantly better performance characteristics over IORef, STRef and MutVar. This is because value is mutated directly in memory instead of following an extra pointer. Besides better performance there is another consequence of direct mutation, namely the value is always evaluated to normal form when being written into a PVar

data PVar a s

Mutable variable with primitive value.

Since: 0.1.0

type RW = RealWorld

Synonym for RealWorld

Creation

newPVar :: (MonadPrim s m, Prim a) => a -> m (PVar a s)

Create a mutable variable in unpinned memory (i.e. GC can move it) with an initial value. This is a prefered way to create a mutable variable, since it will not contribute to memory fragmentation. For pinned memory versions see newPinnedPVar and newAlignedPinnedPVar

Since: 0.1.0

withPVarST

Arguments

:: Prim p

=> p
Initial value assigned to the mutable variable

-> (forall s. PVar p s -> ST s a)
Action to run

-> a
Result produced by the ST action

Run an ST action on a mutable variable.

Since: 0.1.0

Mutable Operations

readPVar :: (MonadPrim s m, Prim a) => PVar a s -> m a

Read a value from a mutable variable

Since: 0.1.0

writePVar :: (MonadPrim s m, Prim a) => PVar a s -> a -> m ()

Write a value into a mutable variable

Since: 0.1.0

modifyPVar :: (MonadPrim s m, Prim a) => PVar a s -> (a -> (a, b)) -> m b

Apply a pure function to the contents of a mutable variable. Returns the artifact of computation.

Since: 0.2.0

modifyPVar_ :: (MonadPrim s m, Prim a) => PVar a s -> (a -> a) -> m ()

Apply a pure function to the contents of a mutable variable.

Since: 0.1.0

fetchModifyPVar :: (MonadPrim s m, Prim a) => PVar a s -> (a -> a) -> m a

Apply a pure function to the contents of a mutable variable. Returns the old value.

Since: 0.2.0

modifyFetchPVar :: (MonadPrim s m, Prim a) => PVar a s -> (a -> a) -> m a

Apply a pure function to the contents of a mutable variable. Returns the new value.

Since: 0.2.0

modifyPVarM :: (MonadPrim s m, Prim a) => PVar a s -> (a -> m (a, b)) -> m b

Apply a monadic action to the contents of a mutable variable. Returns the artifact of computation.

Since: 0.2.0

modifyPVarM_ :: (MonadPrim s m, Prim a) => PVar a s -> (a -> m a) -> m ()

Apply a monadic action to the contents of a mutable variable.

Since: 0.1.0

fetchModifyPVarM :: (MonadPrim s m, Prim a) => PVar a s -> (a -> m a) -> m a

Apply a monadic action to the contents of a mutable variable. Returns the old value.

Since: 0.2.0

modifyFetchPVarM :: (MonadPrim s m, Prim a) => PVar a s -> (a -> m a) -> m a

Apply a monadic action to the contents of a mutable variable. Returns the new value.

Since: 0.2.0

swapPVars_ :: (MonadPrim s m, Prim a) => PVar a s -> PVar a s -> m ()

Swap contents of two mutable variables.

Since: 0.1.0

swapPVars :: (MonadPrim s m, Prim a) => PVar a s -> PVar a s -> m (a, a)

Swap contents of two mutable variables. Returns their old values.

Since: 0.1.0

copyPVar

Arguments

:: (MonadPrim s m, Prim a)

=> PVar a s
Source variable

-> PVar a s
Destination variable

-> m ()

Copy contents of one mutable variable PVar into another

Since: 0.1.0

sizeOfPVar :: Prim a => PVar a s -> Int

Size in bytes of a value stored inside the mutable variable. PVar itself is neither accessed nor evaluated.

Since: 0.1.0

alignmentPVar :: Prim a => PVar a s -> Int

Alignment in bytes of the value stored inside of the mutable variable. PVar itself is neither accessed nor evaluated.

Since: 0.1.0

Pinned memory

In theory it is unsafe to mix Storable and Prim operations on the same chunk of memory, because some instances can have different memory layouts for the same type. This is highly uncommon in practice and if you are intermixing the two concepts together you probably already know what you are doing.

newPinnedPVar :: (MonadPrim s m, Prim a) => a -> m (PVar a s)

Create a mutable variable in pinned memory with an initial value.

Since: 0.1.0

newAlignedPinnedPVar :: (MonadPrim s m, Prim a) => a -> m (PVar a s)

Create a mutable variable in pinned memory with an initial value and aligned according to its alignment

Since: 0.1.0

withPtrPVar :: (MonadPrim s m, Prim a) => PVar a n -> (Ptr a -> m b) -> m (Maybe b)

Apply an action to the Ptr that references the mutable variable, but only if it is backed by pinned memory, cause otherwise it would be unsafe.

Since: 0.1.0

withStorablePVar

Arguments

:: (MonadPrim s m, Storable a)

=> a
Initial value

-> (PVar a s -> Ptr a -> m b)
Action to run

-> m b

Apply an action to the newly allocated PVar and to the Ptr that references it. Memory allocated with number of bytes specified by sizeOf a is allocated and pinned, therefore it is safe to operate directly with the pointer as well as over FFI. Returning the pointer from the supplied action would be very unsafe, therefore return the PVar if you still need it afterwards, garbage collector will cleanup the memory when it is no longer needed.

Since: 0.1.0

withAlignedStorablePVar

Arguments

:: (MonadPrim s m, Storable a)

=> a
Initial value

-> (PVar a s -> Ptr a -> m b)
Action to run

-> m b

Same withStorablePVar, except memory is aligned according to alignment.

Since: 0.1.0

copyPVarToPtr :: (MonadPrim s m, Prim a) => PVar a s -> Ptr a -> m ()

Copy contents of a mutable variable PVar into a pointer Ptr

Since: 0.1.0

toForeignPtrPVar :: PVar a s -> Maybe (ForeignPtr a)

Convert PVar into a ForeignPtr, but only if it is backed by pinned memory.

Since: 0.1.0

isPinnedPVar :: PVar a s -> Bool

Check if PVar is backed by pinned memory or not

Since: 0.1.0

peekPrim :: (Storable a, MonadPrim s m) => Ptr a -> m a

Use Storable reading functionality inside the PrimMonad.

Since: 0.1.0

pokePrim :: (Storable a, MonadPrim s m) => Ptr a -> a -> m ()

Use Storable wrting functionality inside the PrimMonad.

Since: 0.1.0

Atomic operations

atomicModifyIntPVar :: MonadPrim s m => PVar Int s -> (Int -> (Int, a)) -> m a

Apply a function to an integer element of a PVar atomically. Implies a full memory barrier.

Since: 0.1.0

atomicModifyIntPVar_ :: MonadPrim s m => PVar Int s -> (Int -> Int) -> m ()

Apply a function to an integer element of a PVar atomically. Returns the old value. Implies a full memory barrier.

Since: 0.1.0

atomicFetchModifyIntPVar :: MonadPrim s m => PVar Int s -> (Int -> Int) -> m Int

Apply a function to an integer element of a PVar atomically. Implies a full memory barrier. Returns the new value.

Since: 0.2.0

atomicModifyFetchIntPVar :: MonadPrim s m => PVar Int s -> (Int -> Int) -> m Int

Apply a function to an integer element of a PVar atomically. Implies a full memory barrier. Returns the new value.

Since: 0.2.0

atomicReadIntPVar :: MonadPrim s m => PVar Int s -> m Int

Read a value from PVar atomically. Implies a full memory barrier.

Since: 0.1.0

atomicWriteIntPVar :: MonadPrim s m => PVar Int s -> Int -> m ()

Write a value into an PVar atomically. Implies a full memory barrier.

Since: 0.1.0

casIntPVar

Arguments

:: MonadPrim s m

=> PVar Int s
Variable to mutate

-> Int
Old expected value

-> Int
New value

-> m Int
Old actual value

Compare and swap. This is also a function that is used to implement atomicModifyIntPVar. Implies a full memory barrier.

Since: 0.1.0

atomicAddIntPVar :: MonadPrim s m => PVar Int s -> Int -> m Int

Add two numbers, corresponds to (+) done atomically. Returns the previous value of the mutable variable. Implies a full memory barrier.

Since: 0.1.0

atomicSubIntPVar :: MonadPrim s m => PVar Int s -> Int -> m Int

Subtract two numbers, corresponds to (-) done atomically. Returns the previous value of the mutable variable. Implies a full memory barrier.

Since: 0.1.0

atomicAndIntPVar :: MonadPrim s m => PVar Int s -> Int -> m Int

Binary conjuction (AND), corresponds to (.&.) done atomically. Returns the previous value of the mutable variable. Implies a full memory barrier.

Since: 0.1.0

atomicNandIntPVar :: MonadPrim s m => PVar Int s -> Int -> m Int

Binary negation of conjuction (NAND), corresponds to \x y -> complement (x .&. y) done atomically. Returns the previous value of the mutable variable. Implies a full memory barrier.

Since: 0.1.0

atomicOrIntPVar :: MonadPrim s m => PVar Int s -> Int -> m Int

Binary disjunction (OR), corresponds to (.|.) done atomically. Returns the previous value of the mutable variable. Implies a full memory barrier.

Since: 0.1.0

atomicXorIntPVar :: MonadPrim s m => PVar Int s -> Int -> m Int

Binary exclusive disjunction (XOR), corresponds to xor done atomically. Returns the previous value of the mutable variable. Implies a full memory barrier.

Since: 0.1.0

atomicNotIntPVar :: MonadPrim s m => PVar Int s -> m Int

Binary negation (NOT), corresponds to ones' complement done atomically. Returns the previous value of the mutable variable. Implies a full memory barrier.

Since: 0.1.0

Re-exports

class Prim a

Class of types supporting primitive array operations. This includes interfacing with GC-managed memory (functions suffixed with ByteArray#) and interfacing with unmanaged memory (functions suffixed with Addr#). Endianness is platform-dependent.

Minimal complete definition

sizeOf#, alignment#, indexByteArray#, readByteArray#, writeByteArray#, setByteArray#, indexOffAddr#, readOffAddr#, writeOffAddr#, setOffAddr#

class (PrimMonad m, s ~ PrimState m) => MonadPrim s (m :: Type -> Type)

PrimMonad's state token type can be annoying to handle in constraints. This typeclass lets users (visually) notice PrimState equality constraints less, by witnessing that s ~ PrimState m.

class Monad m => PrimMonad (m :: Type -> Type)

Class of monads which can perform primitive state-transformer actions.

Minimal complete definition

primitive

Associated Types

type PrimState (m :: Type -> Type)

State token type.

data RealWorld

RealWorld is deeply magical. It is primitive, but it is not unlifted (hence ptrArg). We never manipulate values of type RealWorld; it's only used in the type system, to parameterise State#.

sizeOf :: Prim a => a -> Int

Size of values of type a. The argument is not used.

This function has existed since 0.1, but was moved from Primitive to Types in version 0.6.3.0.

alignment :: Prim a => a -> Int

Alignment of values of type a. The argument is not used.

This function has existed since 0.1, but was moved from Primitive to Types in version 0.6.3.0.

data ST s a

The strict ST monad. The ST monad allows for destructive updates, but is escapable (unlike IO). A computation of type ST s a returns a value of type a, and execute in "thread" s. The s parameter is either

an uninstantiated type variable (inside invocations of runST), or

RealWorld (inside invocations of stToIO).

It serves to keep the internal states of different invocations of runST separate from each other and from invocations of stToIO.

The >>= and >> operations are strict in the state (though not in values stored in the state). For example,

runST (writeSTRef _|_ v >>= f) = _|_

runST :: (forall s. ST s a) -> a

Return the value computed by a state thread. The forall ensures that the internal state used by the ST computation is inaccessible to the rest of the program.

class Storable a where

The member functions of this class facilitate writing values of primitive types to raw memory (which may have been allocated with the above mentioned routines) and reading values from blocks of raw memory. The class, furthermore, includes support for computing the storage requirements and alignment restrictions of storable types.

Memory addresses are represented as values of type Ptr a, for some a which is an instance of class Storable. The type argument to Ptr helps provide some valuable type safety in FFI code (you can't mix pointers of different types without an explicit cast), while helping the Haskell type system figure out which marshalling method is needed for a given pointer.

All marshalling between Haskell and a foreign language ultimately boils down to translating Haskell data structures into the binary representation of a corresponding data structure of the foreign language and vice versa. To code this marshalling in Haskell, it is necessary to manipulate primitive data types stored in unstructured memory blocks. The class Storable facilitates this manipulation on all types for which it is instantiated, which are the standard basic types of Haskell, the fixed size Int types (Int8, Int16, Int32, Int64), the fixed size Word types (Word8, Word16, Word32, Word64), StablePtr, all types from Foreign.C.Types, as well as Ptr.

Minimal complete definition

sizeOf, alignment, (peek | peekElemOff | peekByteOff), (poke | pokeElemOff | pokeByteOff)

Methods

peek :: Ptr a -> IO a

Read a value from the given memory location.

Note that the peek and poke functions might require properly aligned addresses to function correctly. This is architecture dependent; thus, portable code should ensure that when peeking or poking values of some type a, the alignment constraint for a, as given by the function alignment is fulfilled.

poke :: Ptr a -> a -> IO ()

Write the given value to the given memory location. Alignment restrictions might apply; see peek.